+++ Prext/schemes/SLOW.cc
@@ -26,8 +26,8 @@ void SLOW::initialize(int stage)
     if(stage==0){
         traci = TraCIMobilityAccess().get(getParentModule());
 
-        silentTimeThreshold = SimTime(par("silentTimeThreshold"), SIMTIME_S);
-        speedThresholdSqr = par("speedThreshold");
+        silentTimeThreshold = SimTime(par("silentTimeThreshold").longValue(), SIMTIME_S);
+        speedThresholdSqr = par("speedThreshold").doubleValue();
         speedThresholdSqr *= speedThresholdSqr;
 
         slowPeriodStart = 0;
